Add has_audio_playback manifest to PC CIDs identified as requiring it
A recent change in Checkbox[1] broke SRU test runs because it is now
required to define the `has_audio_playback` manifest to be able to run
most audio tests (these tests used to rely on an automatically
discovered "AUDIO" udev category, but it was prone to errors, especially
with some IoT devices).

Without a properly defined `has_audio_playback` manifest, the audio
tests are skippedsimply skipped, resulting in a lot of tests not run at
all on devices running SRU tests.

A list of PC devices falling into this pitfall was made. Using the
`utils/manivest-update.py` script, update (or create) the manifest to
add the `has_audio_playback`.

It is assumed that all these devices have some way of rendering audio,
as they are consumer-facing PC devices with an audio interface (laptops,
desktops, etc.).

[1] https://github.com/canonical/checkbox/pull/1283