Set has_audio_capture manifest to true for laptops (3)
Laptops come with an integrated webcam and microphone. Therefore, they
have the ability to record sounds by default.

This commit sets the `has_audio_capture` manifest entry to `true` for
these devices.

The list of devices comes from the C3 Monitoring page[1].

This commit deals with manifest.json files for devices that were not
recorded in this repository. The default manifest entries mentioned in
[2] have been added (and edited to match the laptop specifications), and
the `has_audio_capture` entry was added as well and set to `true`.

Fix LM-1657

[1]: https://certification.canonical.com/monitoring/
[2]:
https://docs.google.com/document/d/1UuwXcBzN_EUnZzLcO5y2Ef30jOiW_hvV1doRy3XyKDc/edit?pli=1#heading=h.hpbi5teoupc2